Completa il porting a 64 bit di sum-it: build integrale riuscita
Sistema gli ultimi file rimasti con lo stesso pattern già usato per
Formula/Excel/Huffman: variabili long/ulong di stile BeOS R5 allargate
o ristrette a int32/uint32/type_code/status_t per matchare le firme
moderne di Haiku (BMessage::FindInt32/GetInfo, atomic_add,
wait_for_thread, BTextView::GetSelection), piu' un cast esplicito per
un overload ambiguo di LoadCursor e un ultimo isnan mancante nel
namespace std.

Con questi fix il binario OpenSum-It compila e si linka per intero
senza errori, ed e' stato avviato con successo su Haiku hrev59800
senza crash.

+++ legacy/opensumit/sum-it/Source/main/Cell-UI/CellView.drag.cpp
@@ -184,7 +184,7 @@ void CCellView::MouseMoved(BPoint point, ulong transit, const BMessage *message)
 		{
 			if (fCurrentCursor != curPlus)
 			{
-				be_app->SetCursor(gResourceManager.LoadCursor(0L));
+				be_app->SetCursor(gResourceManager.LoadCursor((int32)0));
 				fCurrentCursor = curPlus;
 			}
 		}
