+++ legacy/opensumit/sum-it/Source/main/Utils/Utils.cpp
@@ -205,7 +205,7 @@ bool IsOptionalClick(BMessage *msg)
 {
 	try
 	{
-		long buttons, modifiers;
+		int32 buttons, modifiers;
 		FailOSErr(msg->FindInt32("buttons", &buttons), errMessageMissing);
 		FailOSErr(msg->FindInt32("modifiers", &modifiers), errMessageMissing);
 		return buttons & B_SECONDARY_MOUSE_BUTTON || modifiers & B_CONTROL_KEY;
